Clean up XVIZ Settings (#197)
- Replace `TimeResolution` with `PLAYBACK_FRAME_RATE`
- Add `TIMESTAMP_FORMAT` config
- Remove `PRIMITIVE_SETTINGS`
