+++ docs/api-reference/xviz-configuration.md
@@ -10,6 +10,7 @@ Constants:
 - `PRIMARY_POSE_STREAM` (String) - Name of the vehicle pose stream. Default `vehicle-pose`.
 - `OBJECT_STREAM` (String) - Name of the object stream. Default `objects`.
 - `STREAM_BLACKLIST` (Array) - Names of the streams to block. Default empty.
+- `TIMESTAMP_FORMAT` (String) - `milliseconds` or `seconds`. Default `milliseconds`.
 
 Parser plugins:
 
@@ -24,10 +25,8 @@ Returns the current XVIZ config.
 
 Sets the XVIZ settings. The default settings are:
 
-- `hiTimeResolution` (Number) - Update pose and lightweight geometry up to 60Hz. Default `1 / 60`.
-- `loTimeResolution` (Number) - Throttle expensive geometry updates. Default `1 / 10`
-- `interpolateVehiclePose` (Boolean) - Injects interpolated vehicle pose datums. Default `false`.
-- `pathDistanceThreshold` (Number) - Filters out close vertices (work around for PathLayer issue) Default `0.1`.
+- `TIME_WINDOW` (Number) - The maximum time difference allowed for a state update to be associated with a given timestamp. Increase this number to ensure that we cover a sufficient time window for any available data. Default `400`.
+- `PLAYBACK_FRAME_RATE` (Number) - The number of frames to generate per second. Default `10`.
 
 
 ### getXVIZSettings(config)
