+++ modules/parser/src/parsers/xviz-primitives-v2.js
@@ -1,6 +1,6 @@
 /* global URL, Blob */
 import {filterVertices} from './filter-vertices';
-import {PRIMITIVE_CAT} from './parse-xviz-stream';
+import {PRIMITIVE_CAT} from './parse-xviz-primitive';
 import base64js from 'base64-js';
 
 function aliasId(primitive) {
