+++ test/modules/parser/synchronizers/log-synchronizer.spec.js
@@ -1,4 +1,4 @@
-import {LogSynchronizer} from '@xviz/parser';
+import {LogSynchronizer, setXVIZSettings} from '@xviz/parser';
 import tape from 'tape-catch';
 
 // xviz data uses snake_case
@@ -44,6 +44,7 @@ tape('LogSynchronizer#setTime', t => {
 });
 
 tape('LogSynchronizer#getData', t => {
+  setXVIZSettings({TIME_WINDOW: 0.4});
   const logSynchronizer = new LogSynchronizer(LOG_START_TIME, LOGS);
 
   for (const tc of TEST_CASES) {
